Check fopen results

+++ src/services/Optimize.php
@@ -569,9 +569,9 @@ class Optimize extends Component
                         }
                         try {
                             $variantPath = $asset->getFolder()->path.$assetTransforms->getTransformSubpath(
-                                    $asset,
-                                    $transformIndex
-                                );
+                                $asset,
+                                $transformIndex
+                            );
                         } catch (InvalidConfigException $e) {
                             $variantPath = '';
                             Craft::error(
@@ -610,21 +610,21 @@ class Optimize extends Component
     ) {
         // If the image variant creation succeeded, copy it into place
         if (!empty($outputPath) && is_file($outputPath)) {
-            $volume = null;
             // Figure out the resulting path for the image variant
             try {
                 $volume = $asset->getVolume();
             } catch (InvalidConfigException $e) {
+                $volume = null;
                 Craft::error(
                     'Asset volume error: '.$e->getMessage(),
                     __METHOD__
                 );
             }
             $assetTransforms = Craft::$app->getAssetTransforms();
-            $transformPath = '';
             try {
                 $transformPath = $asset->getFolder()->path.$assetTransforms->getTransformSubpath($asset, $index);
             } catch (InvalidConfigException $e) {
+                $transformPath = '';
                 Craft::error(
                     'Error getting asset folder: '.$e->getMessage(),
                     __METHOD__
@@ -646,19 +646,20 @@ class Optimize extends Component
 
             clearstatcache(true, $outputPath);
             $stream = @fopen($outputPath, 'rb');
+            if ($stream !== false) {
+                // Now create it
+                try {
+                    $volume->createFileByStream($variantPath, $stream, []);
+                } catch (VolumeException $e) {
+                    Craft::error(
+                        Craft::t('image-optimize', 'Failed to create image variant at: ')
+                        .$outputPath,
+                        __METHOD__
+                    );
+                }
 
-            // Now create it
-            try {
-                $volume->createFileByStream($variantPath, $stream, []);
-            } catch (VolumeException $e) {
-                Craft::error(
-                    Craft::t('image-optimize', 'Failed to create image variant at: ')
-                    .$outputPath,
-                    __METHOD__
-                );
+                FileHelper::unlink($outputPath);
             }
-
-            FileHelper::unlink($outputPath);
         } else {
             Craft::error(
                 Craft::t('image-optimize', 'Failed to create image variant at: ')
